Икапсуляция PathFinder завершена
