+++ src/command/Control.hpp
@@ -131,7 +131,7 @@ public:
         }
         
         if (params.command == "PlotPath") {
-            gnuplotInterface.plotPath(path, p, params.filename, params, pathFinder, config.vehicleRadius);
+            gnuplotInterface.plotPath(path, p, params.filename, params, config.vehicleRadius);
             logOperation(LogLevel::Info, std::string("PlotPath"), std::string("file: ") + params.filename);
         }
 
@@ -212,7 +212,7 @@ public:
         if (params.command == "triangulate") {
             clusterCenters = clusterService.getClusterCenters(componenti, config);  
             lastTriangulation = triangulator.bowyerWatson(clusterCenters);
-            voronoi.buildFromDelaunay(lastTriangulation, pathFinder, p, voronoiEdges);
+            voronoi.buildFromDelaunay(lastTriangulation, p, voronoiEdges);
             logOperation(LogLevel::Info, std::string("triangulate"), 
                 std::string("clusters=") + std::to_string(clusterCenters.size()) + 
                 ", triangles=" + std::to_string(lastTriangulation.size()));
@@ -248,14 +248,14 @@ public:
         if (params.command == "Plot3DPath") {
             PointD start(params.startPointX, params.startPointY);
             PointD end(params.endPointX, params.endPointY);
-            gnuplotInterface.plot3DPath(path, p, params.filename, start, end, pathFinder, config.vehicleRadius);
+            gnuplotInterface.plot3DPath(path, p, params.filename, start, end, config.vehicleRadius);
             logOperation(LogLevel::Info, std::string("Plot3DPath"), std::string("file: ") + params.filename);
         }
         
         if (params.command == "plotInteractive3DPath") {
             PointD start(params.startPointX, params.startPointY);
             PointD end(params.endPointX, params.endPointY);
-            gnuplotInterface.plotInteractive3DPath(path, p, start, end, pathFinder, config.vehicleRadius);
+            gnuplotInterface.plotInteractive3DPath(path, p, start, end, config.vehicleRadius);
             logOperation(LogLevel::Info, std::string("plotInteractive3DPath"));
         }
     }
